requirements.txt
pybids==0.6.3

